Moved controllers to another folder
Created folder controllers and moved users.js and contacts.js to there to keep routers cleaner
